Add files via upload
Aspiring to be a dedicated Full stack developer and a passion for web development. Seeking a challenging position in a reputable organization to expand and utilize my learning, skills and knowledge. Eager to embark on a journey of learning and skill-building in order to create exciting and innovative projects. Constantly learning and adapting to the latest technologies, with a focus on delivering high-quality solutions. Contributes to successful teamwork and innovative development.